+++ zcash_primitives/src/sapling/address.rs
@@ -9,8 +9,9 @@ use super::{
 ///
 /// # Invariants
 ///
-/// `pk_d` is guaranteed to be prime-order (i.e. in the prime-order subgroup of Jubjub,
-/// and not the identity).
+/// - `diversifier` is guaranteed to be valid for Sapling (only 50% of diversifiers are).
+/// - `pk_d` is guaranteed to be prime-order (i.e. in the prime-order subgroup of Jubjub,
+///  and not the identity).
 #[derive(Clone, Debug)]
 pub struct PaymentAddress {
     pk_d: jubjub::SubgroupPoint,
@@ -28,8 +29,14 @@ impl Eq for PaymentAddress {}
 impl PaymentAddress {
     /// Constructs a PaymentAddress from a diversifier and a Jubjub point.
     ///
-    /// Returns None if `pk_d` is the identity.
+    /// Returns None if `diversifier` is not valid for Sapling, or `pk_d` is the identity.
+    /// Note that we cannot verify in this constructor that `pk_d` is derived from
+    /// `diversifier`, so addresses for which these values have no known relationship
+    /// (and therefore no-one can receive funds at them) can still be constructed.
     pub fn from_parts(diversifier: Diversifier, pk_d: jubjub::SubgroupPoint) -> Option<Self> {
+        // Check that the diversifier is valid
+        diversifier.g_d()?;
+
         if pk_d.is_identity().into() {
             None
         } else {
@@ -39,7 +46,9 @@ impl PaymentAddress {
 
     /// Constructs a PaymentAddress from a diversifier and a Jubjub point.
     ///
-    /// Only for test code, as this explicitly bypasses the invariant.
+    /// Only for test code, as this explicitly bypasses the invariant. [`Self::g_d`] will
+    /// panic on a `PaymentAddress` that has been constructed by passing an invalid
+    /// diversifier to this method.
     #[cfg(test)]
     pub(crate) fn from_parts_unchecked(
         diversifier: Diversifier,
@@ -55,11 +64,10 @@ impl PaymentAddress {
             tmp.copy_from_slice(&bytes[0..11]);
             Diversifier(tmp)
         };
-        // Check that the diversifier is valid
-        diversifier.g_d()?;
 
         let pk_d = jubjub::SubgroupPoint::from_bytes(bytes[11..43].try_into().unwrap());
         if pk_d.is_some().into() {
+            // The remaining invariants are checked here.
             PaymentAddress::from_parts(diversifier, pk_d.unwrap())
         } else {
             None
@@ -84,17 +92,17 @@ impl PaymentAddress {
         &self.pk_d
     }
 
-    pub fn g_d(&self) -> Option<jubjub::SubgroupPoint> {
-        self.diversifier.g_d()
+    pub(crate) fn g_d(&self) -> jubjub::SubgroupPoint {
+        self.diversifier.g_d().expect("checked at construction")
     }
 
-    pub fn create_note(&self, value: u64, rseed: Rseed) -> Option<Note> {
-        self.g_d().map(|g_d| Note {
+    pub fn create_note(&self, value: u64, rseed: Rseed) -> Note {
+        Note {
             value,
             rseed,
-            g_d,
+            g_d: self.g_d(),
             pk_d: self.pk_d,
-        })
+        }
     }
 }