+++ zcash_primitives/src/sapling/note.rs
@@ -133,7 +133,7 @@ pub(super) mod testing {
         ) -> Note {
             Note {
                 value: value.inner(),
-                g_d: addr.g_d().unwrap(), // this unwrap is safe because arb_payment_address always generates an address with a valid g_d
+                g_d: addr.g_d(),
                 pk_d: *addr.pk_d(),
                 rseed
             }